add Activity Connection add certification add connection add inscription add Activity Parameters fix bugs add self,join

+++ app/src/main/java/io/ucoin/app/model/UcoinBlock.java
@@ -17,7 +17,7 @@ public interface UcoinBlock extends SqlRow {
 
     Long medianTime();
 
-    Long dividend();
+    BigInteger dividend();
 
     BigInteger monetaryMass();
